requirements.txt
langgraph~=1.0.4
langchain-core>=1.1.1
langchain-openai>=1.1.0
python-dotenv~=0.21.0
dotenv~=0.9.9
config~=0.5.1
python-toon~=0.1.3